Cbewsl with edit and delete

+++ app/components/Dashboard/risk_assesment/save_hazard_data.js
@@ -9,6 +9,8 @@ export default class SaveHazardData extends Component {
         super(props);
         this.state = {
             hazard_data_id: 0,
+            local_storage_id: 0,
+            sync_status: 0,
             hazard: "",
             speed_of_onset: "",
             early_warning: "",
@@ -19,9 +21,12 @@ export default class SaveHazardData extends Component {
     componentWillMount() {
         const { navigation } = this.props;
         const data = navigation.getParam("data", "none");
+        console.log(data)
         if (data != "none") {
             this.setState({
                 hazard_data_id: data.hazard_data_id,
+                local_storage_id: data.local_storage_id,
+                sync_status: data.sync_status,
                 hazard: data.hazard,
                 speed_of_onset: data.speed_of_onset,
                 early_warning: data.early_warning,
@@ -30,6 +35,8 @@ export default class SaveHazardData extends Component {
         } else {
             this.setState({
                 hazard_data_id: 0,
+                local_storage_id: 0,
+                sync_status: 0,
                 hazard: "",
                 speed_of_onset: "",
                 early_warning: "",
@@ -40,6 +47,8 @@ export default class SaveHazardData extends Component {
 
     saveHazardData() {
         const { hazard_data_id,
+            local_storage_id,
+            sync_status,
             hazard,
             speed_of_onset,
             early_warning,
@@ -53,6 +62,8 @@ export default class SaveHazardData extends Component {
             },
             body: JSON.stringify({
                 hazard_data_id: hazard_data_id,
+                local_storage_id: local_storage_id,
+                sync_status: sync_status,
                 hazard: hazard,
                 speed_of_onset: speed_of_onset,
                 early_warning: early_warning,
@@ -62,6 +73,42 @@ export default class SaveHazardData extends Component {
             .then((responseJson) => {
                 if (responseJson.status == true) {
                     ToastAndroid.show(responseJson.message, ToastAndroid.SHORT);
+                    let data_container = Storage.getItem('RiskAssessmentHazardData')
+                    let updated_data = []
+                    data = {
+                        hazard_data_id: hazard_data_id,
+                        local_storage_id: 1,
+                        sync_status: 3,
+                        hazard: hazard,
+                        speed_of_onset: speed_of_onset,
+                        early_warning: early_warning,
+                        impact: impact
+                    }
+                    data_container.then(response => {
+                        if (response == null) {
+                            Storage.removeItem("RiskAssessmentHazardData")
+                            Storage.setItem("RiskAssessmentHazardData", [data])
+                        } else {
+                            let temp = response
+                            temp.push(data)
+                            let counter = 0
+                            temp.forEach((value) => {
+                                counter += 1
+                                updated_data.push({
+                                    hazard_data_id: value.hazard_data_id,
+                                    local_storage_id: counter,
+                                    sync_status: 3,
+                                    hazard: hazard,
+                                    speed_of_onset: value.speed_of_onset,
+                                    early_warning: value.early_warning,
+                                    impact: value.impact
+                                })
+                            });
+                            Storage.removeItem("RiskAssessmentHazardData")
+                            Storage.setItem("RiskAssessmentHazardData", updated_data)
+                        }
+                    });
+
                     this.props.navigation.navigate('modify_hazard_data');
                 } else {
                     ToastAndroid.show(responseJson.message, ToastAndroid.SHORT);
@@ -70,6 +117,8 @@ export default class SaveHazardData extends Component {
             .catch((error) => {
                 data = {
                     hazard_data_id: hazard_data_id,
+                    local_storage_id: local_storage_id,
+                    sync_status: 1,
                     hazard: hazard,
                     speed_of_onset: speed_of_onset,
                     early_warning: early_warning,
@@ -77,14 +126,60 @@ export default class SaveHazardData extends Component {
                 }
                 let offline_data = Storage.getItem("RiskAssessmentHazardData");
                 offline_data.then(response => {
-                    if (response == null) {
-                        Storage.removeItem("RiskAssessmentHazardData")
-                        Storage.setItem("RiskAssessmentHazardData", [data])
+                    if (local_storage_id == 0) {
+                        if (response == null) {
+                            Storage.removeItem("RiskAssessmentHazardData")
+                            Storage.setItem("RiskAssessmentHazardData", [data])
+                        } else {
+                            let temp = response
+                            temp.push(data)
+                            let updated_data = []
+                            let counter = 0
+                            temp.forEach((value) => {
+                                counter += 1
+                                updated_data.push({
+                                    hazard_data_id: value.hazard_data_id,
+                                    local_storage_id: counter,
+                                    sync_status: value.sync_status,
+                                    hazard: value.hazard,
+                                    speed_of_onset: value.speed_of_onset,
+                                    early_warning: value.early_warning,
+                                    impact: value.impact
+                                })
+                            });
+                            Storage.removeItem("RiskAssessmentHazardData")
+                            Storage.setItem("RiskAssessmentHazardData", updated_data)
+                        }
                     } else {
                         let temp = response
-                        temp.push(data)
+                        let updated_data = []
+                        let counter = 0
+                        temp.forEach((value) => {
+                            counter += 1
+                            if (local_storage_id == value.local_storage_id) {
+                                updated_data.push({
+                                    hazard_data_id: hazard_data_id,
+                                    local_storage_id: counter,
+                                    sync_status: 2,
+                                    hazard: hazard,
+                                    speed_of_onset: speed_of_onset,
+                                    early_warning: early_warning,
+                                    impact: impact
+                                })
+                            } else {
+                                updated_data.push({
+                                    hazard_data_id: value.hazard_data_id,
+                                    local_storage_id: counter,
+                                    sync_status: value.sync_status,
+                                    hazard: value.hazard,
+                                    speed_of_onset: value.speed_of_onset,
+                                    early_warning: value.early_warning,
+                                    impact: value.impact
+                                })
+                            }
+                        });
                         Storage.removeItem("RiskAssessmentHazardData")
-                        Storage.setItem("RiskAssessmentHazardData", temp)
+                        Storage.setItem("RiskAssessmentHazardData", updated_data)
                     }
                 })
                 this.props.navigation.navigate('modify_hazard_data');
